2 Big Changes in Event and the Main File
RunEvent is now two different methods so that backgrounds draw in the
main file before running the whole event.

The main file now goes to the correct screen after stats, running the
game. Changed the main game loop by adding in food calculations and
various other things I previously implemented wrong.

2 things need to be fixed:
-Next button increments room twice for some reason, although I worked
around it. It still looks bad in the code
-The background images don't actually load on screen, although the code
runs it correctly. Might be a problem with the image files themselves,
but once we fix it the code should already be working correctly.